requirements.txt
# Func関連
azure-functions==1.21.3
cachetools==5.5.0
# certifi==2024.8.30
# charset-normalizer==3.4.0
# idna==3.10
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
# requests==2.32.3
rsa==4.9
# urllib3==2.2.3

# GA関連
google-analytics-data==0.18.15
google-api-core==2.24.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
python-dotenv==1.0.1

# DB関連
azure-core==1.26.4
azure-cosmos==4.3.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
idna==3.4
pip-upgrade==0.0.6
platformdirs==3.2.0
pur==7.1.0
requests==2.28.2
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.22.0
virtualenv-clone==0.5.7
